Move free methods in assimp_loader to class methods
